# Nueva carpeta (2)/microservice-credit/src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Credit Microservice
  description: API to manage credits.
  version: 1.0.0

servers:
  - url: http://localhost:8092

paths:

  /credits:
    post:
      tags:
        - credit
      summary: Create a new credit card
      operationId: createCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditInput'
      responses:
        201:
          description: Credit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
        400:
          description: Invalid input
  /credits/{id}:
    get:
      tags:
        - credit
      summary: Find credit by ID
      operationId: findCreditById
      parameters:
        - in: path
          name: id
          description: ID of credit to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
        400:
          description: Invalid ID supplied
        404:
          description: Credit not found
  /credits/client/{id}:
    get:
      tags:
        - credit
      summary: Find all credits of a client
      operationId: getCreditsOfClient
      parameters:
        - in: path
          name: id
          description: ID of client to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credit'
        400:
          description: Invalid ID supplied
        404:
          description: Account not found

components:
  schemas:
    Credit:
      type: object
      properties:
        id:
          type: string
          example: 1
        cardType:
          type: string
          enum: [personal, business]
        creditLimit:
          type: number
          format: float
          example: 100.0
        outstandingBalance:
          type: number
          format: float
          example: 10.0
        clientId:
          type: string
          example: 1
    CreditInput:
      type: object
      properties:
        cardType:
          type: string
          enum: [personal, business]
        creditLimit:
          type: number
          format: float
          example: 100.0
        clientId:
          type: string
          example: 1

tags:
  - name: credit
    description: Operations about credits